Improve reference to external models
* Modify the part that used the internally copied models to use the external models
* Remove the copied models from the external subsystems
