--- pkg/api/rest/model/onprem/network/dns.go
@@ -1,6 +1,0 @@
-package network
-
-// DNS TODO
-type DNS struct {
-	DNSServer []string `json:"dns_server"` // IPv4 or IPv6 DNS Server Addresses
-}
