--- pkg/api/rest/model/onprem/network/firewall.go
@@ -1,12 +1,0 @@
-package network
-
-type FirewallRule struct {
-	Priority  uint   `json:"priority"` // Lower has higher priority
-	Src       string `json:"src"`
-	Dst       string `json:"dst"`
-	SrcPorts  string `json:"src_ports"`
-	DstPorts  string `json:"dst_ports"`
-	Protocol  string `json:"protocol"`  // TCP, UDP, ICMP
-	Direction string `json:"direction"` // inbound, outbound
-	Action    string `json:"action"`    // allow, deny
-}
