--- pkg/api/rest/model/onprem/network/nic.go
@@ -1,9 +1,0 @@
-package network
-
-type NIC struct {
-	Interface  string   `json:"interface"`
-	Address    []string `json:"address"`
-	Gateway    []string `json:"gateway"`
-	MACAddress string   `json:"mac_address"`
-	MTU        int      `json:"mtu"`
-}
