--- pkg/api/rest/model/onprem/network/route.go
@@ -1,7 +1,0 @@
-package network
-
-type Route struct {
-	Destination string `json:"destination"`
-	Netmask     string `json:"netmask"`
-	NextHop     string `json:"next_hop"`
-}
