--- pkg/api/rest/model/onprem/software/docker.go
@@ -1,8 +1,0 @@
-package software
-
-import "github.com/docker/docker/api/types"
-
-type Docker struct {
-	Containers []types.Container
-	//Images     []types.ImageMetadata
-}
