--- pkg/api/rest/model/onprem/software/podman.go
@@ -1,8 +1,0 @@
-package software
-
-import "github.com/docker/docker/api/types"
-
-type Podman struct {
-	Containers []types.Container
-	//Images     []types.ImageMetadata
-}
