--- pkg/api/rest/model/onprem/software/software.go
@@ -1,8 +1,0 @@
-package software
-
-type Software struct {
-	DEB    []DEB  `json:"deb"`
-	RPM    []RPM  `json:"rpm"`
-	Docker Docker `json:"docker"`
-	Podman Podman `json:"podman"`
-}
